Optimize USRP sc16 streaming and raise test rate to 64 kbps

+++ UAV_Link_Sim/usrp_driver.h
@@ -9,6 +9,7 @@
 #include <complex>
 #include <string>
 #include <cstddef>
+#include <cstdint>
 
 using Complex = std::complex<double>;
 using VecComplex = std::vector<Complex>;
@@ -67,7 +68,7 @@ private:
     std::atomic<int> rx_overflow_count_{ 0 };
     std::thread rx_thread_;
 
-    std::vector<std::complex<float>> rx_buffer_raw_;
+    std::vector<std::complex<int16_t>> rx_buffer_raw_;
     std::string effective_device_args_;
     double actual_master_clock_rate_ = 0.0;
     double actual_rx_rate_ = 0.0;
